% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koppen_functions.R
\name{calculate_pwinter}
\alias{calculate_pwinter}
\title{Internal function to determine the summer precipitation months}
\usage{
calculate_pwinter(prec, lat)
}
\arguments{
\item{prec}{A vector with precipitation for each month}

\item{lat}{The latitude}
}
\value{
A vector with the precipitation in the "summer" month as defined for Köppen climate classification
}
\description{
Internal function to determine the summer precipitation months
}
