% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_climate_koppen.R
\name{calculate_threshold}
\alias{calculate_threshold}
\title{Internal function to determine the Köppen threshold number}
\usage{
calculate_threshold(tavg, ptot, psummer, pwinter)
}
\arguments{
\item{tavg}{The annual average temperature}

\item{ptot}{The annual total precipitation}

\item{psummer}{A vector with the precipitation in summer months}

\item{pwinter}{A vector with the precipitation in winter months}
}
\value{
The threshold number for Köppen evaporation (for B climates)
}
\description{
Internal function to determine the Köppen threshold number
}
