% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_carto.R
\name{carto_ne_get_boundbox}
\alias{carto_ne_get_boundbox}
\title{Return NaturalEarth Bound box}
\usage{
carto_ne_get_boundbox()
}
\value{
A simple feature
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
