% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_cut.R
\name{chop_relabel_dash}
\alias{chop_relabel_dash}
\title{Reformat santoku chop dash labels.}
\usage{
chop_relabel_dash(x)
}
\arguments{
\item{x}{A string vector of labels}
}
\value{
A string vector with clean labels
}
\description{
Reformat santoku chop dash labels.
}
