% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cut.R
\name{cut_relabel}
\alias{cut_relabel}
\title{Reformat cut labels.}
\usage{
cut_relabel(x)
}
\arguments{
\item{x}{A string vector of labels}
}
\value{
A string vector with clean labels
}
\description{
Reformat cut labels.
}
