% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_climate_koppen.R
\name{determine_koppen_climate}
\alias{determine_koppen_climate}
\title{Function: Determine the Köppen climate based on temperature,
precipitation and the latitude.}
\usage{
determine_koppen_climate(temp, prec, lat)
}
\arguments{
\item{temp}{A vector with temperatures for each month}

\item{prec}{A vector with precipitations for each month}

\item{lat}{The latitude}
}
\value{
A string with the Köppen climate determination
}
\description{
Function: Determine the Köppen climate based on temperature,
precipitation and the latitude.
}
