% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_equalearth_cutout}
\alias{df_cartography_equalearth_cutout}
\title{DF Cartography: Equal Earth cutout}
\usage{
df_cartography_equalearth_cutout()
}
\value{
A sf object
}
\description{
Generate Equal Earth special cutout shape
}
\keyword{internal}
