% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_equalearth_outline}
\alias{df_cartography_equalearth_outline}
\title{DF Cartography: Equal Earth outline}
\usage{
df_cartography_equalearth_outline()
}
\value{
A sf object
}
\description{
Generate Equal Earth special outline shape
}
\keyword{internal}
