% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_goodehomolosine_cutout}
\alias{df_cartography_goodehomolosine_cutout}
\title{Goode Homolosine cutout}
\usage{
df_cartography_goodehomolosine_cutout()
}
\value{
A sf object
}
\description{
Generate Goode Homolosine special cutout shape
}
\keyword{internal}
