% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_nonrectangular_projections}
\alias{df_cartography_nonrectangular_projections}
\title{DF Cartography: Patchwork of pseudo projections}
\usage{
df_cartography_nonrectangular_projections()
}
\value{
A patchwork object
}
\description{
A patchwork composition of Goode Homolosine, Robinson,
Winkel-Tripel and Equal Earth projections.
}
