% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_project_equalearth}
\alias{df_cartography_project_equalearth}
\title{DF Cartography: Equal Earth projection}
\usage{
df_cartography_project_equalearth()
}
\value{
A ggplot2 object
}
\description{
ggplot2 map of the world in Equal Earth projection
}
