% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_project_equirectangular}
\alias{df_cartography_project_equirectangular}
\title{DF Cartography: Equirectangular projection}
\usage{
df_cartography_project_equirectangular()
}
\value{
A ggplot2 object
}
\description{
ggplot2 map of the world in Equirectangular projection
}
