% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_project_gallpeters}
\alias{df_cartography_project_gallpeters}
\title{DF Cartography: Gall-Peters projection}
\usage{
df_cartography_project_gallpeters()
}
\value{
A ggplot2 object
}
\description{
ggplot2 map of the world in Gall-Peters projection
}
