% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_project_goodehomolosine}
\alias{df_cartography_project_goodehomolosine}
\title{DF Cartography: Goode Homolosine projection}
\usage{
df_cartography_project_goodehomolosine()
}
\value{
A ggplot2 object
}
\description{
ggplot2 map of the world in Goode Homolosine projection
}
