% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_project_hobodyer}
\alias{df_cartography_project_hobodyer}
\title{DF Cartography: Hobo-Dyer projection}
\usage{
df_cartography_project_hobodyer()
}
\value{
A ggplot2 object
}
\description{
ggplot2 map of the world in Hobo-Dyer projection
}
