% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_project_mercator}
\alias{df_cartography_project_mercator}
\title{DF Cartography: Mercator projection}
\usage{
df_cartography_project_mercator()
}
\value{
A ggplot2 object
}
\description{
ggplot2 map of the world in Mercator projection
}
