% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_project_robinson}
\alias{df_cartography_project_robinson}
\title{DF Cartography: Robinson projection}
\usage{
df_cartography_project_robinson()
}
\value{
A ggplot2 object
}
\description{
ggplot2 map of the world in Robinson projection
}
\examples{
df_cartography_project_robinson()
}
