% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_project_winkeltripel}
\alias{df_cartography_project_winkeltripel}
\title{DF Cartography: Winkel-Tripel projection}
\usage{
df_cartography_project_winkeltripel()
}
\value{
A ggplot2 object
}
\description{
ggplot2 map of the world in Winkel-Tripel projection
}
