% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_projections}
\alias{df_cartography_projections}
\alias{df_cartography_project_equirectangular}
\alias{df_cartography_project_gallpeters}
\alias{df_cartography_project_hobodyer}
\alias{df_cartography_project_mercator}
\alias{df_cartography_project_winkeltripel}
\alias{df_cartography_project_robinson}
\alias{df_cartography_project_goodehomolosine}
\alias{df_cartography_project_equalearth}
\title{Showcase different projections}
\usage{
df_cartography_project_equirectangular()

df_cartography_project_gallpeters()

df_cartography_project_hobodyer()

df_cartography_project_mercator()

df_cartography_project_winkeltripel()

df_cartography_project_robinson()

df_cartography_project_goodehomolosine()

df_cartography_project_equalearth()
}
\value{
A ggplot2 object
}
\description{
These functions showcase different map projections.
}
\details{
The following projections are available:
\itemize{
\item Equi-rectangular: \code{df_cartography_project_equirectangular()}
\item Gall-Peters: \code{df_cartography_project_gallpeters()}
\item Hobo-Dyer: \code{df_cartography_project_hobodyer()}
\item Mercator: \code{df_cartography_project_mercator()}
\item Wikel-Tripel: \code{df_cartography_project_winkeltripel()}
\item Robinson: \code{df_cartography_project_robinson()}
\item Goode Homolosine: \code{df_cartography_project_goodehomolosine()}
\item Equal Earth: \code{df_cartography_project_equalearth()}
}
}
\examples{
df_cartography_project_equirectangular()
df_cartography_project_gallpeters()
df_cartography_project_hobodyer()
df_cartography_project_mercator()
df_cartography_project_winkeltripel()
df_cartography_project_robinson()
df_cartography_project_goodehomolosine()
df_cartography_project_equalearth()
}
