% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_rectangular_projections}
\alias{df_cartography_rectangular_projections}
\title{DF Cartography: Patchwork of rectangular projections}
\usage{
df_cartography_rectangular_projections()
}
\value{
A patchwork object
}
\description{
A patchwork composition of Equirectangular, Mercator,
Gall-Peters and Hobo Dyer projections.
}
