% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_robinson_cutout}
\alias{df_cartography_robinson_cutout}
\title{DF Cartography: Robinson cutout}
\usage{
df_cartography_robinson_cutout()
}
\value{
A sf object
}
\description{
Generate Robinson special cutout shape
}
\keyword{internal}
