% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_df_demography.R
\name{df_demography_graph_world_population}
\alias{df_demography_graph_world_population}
\title{DF Demography: World population patchwork composition}
\usage{
df_demography_graph_world_population(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
World population graph with historical, current and future
numbers as well as growth rates.
}
