% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_df_demography.R
\name{df_demography_graph_world_population_current}
\alias{df_demography_graph_world_population_current}
\title{World population}
\usage{
df_demography_graph_world_population_current(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
World population graph
}
\examples{
df_demography_graph_world_population_current()
}
