% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_df_demography.R
\name{df_demography_graph_world_population_growth}
\alias{df_demography_graph_world_population_growth}
\title{DF Demography: World population growth rate}
\usage{
df_demography_graph_world_population_growth(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
World population growth rate graph
}
