% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_df_demography.R
\name{df_demography_graph_world_population_historical}
\alias{df_demography_graph_world_population_historical}
\title{DF Demography: Historical world population plot}
\usage{
df_demography_graph_world_population_historical(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
World population graph from 10'000 BCE to 1200 CE
}
\examples{
df_demography_graph_world_population_historical()
}
