% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_df_demography.R
\name{df_demography_hc_world_hdi}
\alias{df_demography_hc_world_hdi}
\title{Graph of the Human Development Index}
\usage{
df_demography_hc_world_hdi()
}
\value{
A hicharter graph
}
\description{
Every country of the world with the Human Development Index
from 1990 to 2021.
}
\examples{
df_demography_hc_world_hdi()
}
