% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_demography.R
\name{df_demography_map_hc_world_hdi}
\alias{df_demography_map_hc_world_hdi}
\title{Humand Development Index (HDI) map}
\usage{
df_demography_map_hc_world_hdi(year)
}
\arguments{
\item{year}{The year}
}
\value{
A highcharter map
}
\description{
Humand Development Index (HDI) map
}
\examples{
df_demography_map_hc_world_hdi(2020)
}
