% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_swissmap.R
\name{geo_create_voting_maps}
\alias{geo_create_voting_maps}
\title{Write voting maps for communal and cantonal level to path}
\usage{
geo_create_voting_maps(url)
}
\arguments{
\item{url}{The url where the data comes from}

\item{path}{A path where to write the maps}
}
\description{
Write voting maps for communal and cantonal level to path
}
