% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_world.R
\name{geo_historical_world_map}
\alias{geo_historical_world_map}
\title{Return Historical World Map}
\usage{
geo_historical_world_map(date)
}
\arguments{
\item{date}{A date object}
}
\value{
An sf object
}
\description{
Return Historical World Map
}
