% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_swiss.R
\name{geo_map_swiss_lakes}
\alias{geo_map_swiss_lakes}
\title{Swiss lakes for maps}
\usage{
geo_map_swiss_lakes(fill_color = "skyblue")
}
\arguments{
\item{fill_color}{A color for the lakes (default: skyblue)}
}
\value{
A ggplot 2 layer for Swiss lakes
}
\description{
Swiss lakes for maps
}
