% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_swiss.R
\name{geo_map_swiss_voting}
\alias{geo_map_swiss_voting}
\title{Write voting maps for communal and cantonal level to path}
\usage{
geo_map_swiss_voting(url)
}
\arguments{
\item{url}{The url where the data comes from}

\item{path}{A path where to write the maps}
}
\description{
Write voting maps for communal and cantonal level to path
}
