% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_swiss.R
\name{geo_map_swiss_voting_walk}
\alias{geo_map_swiss_voting_walk}
\title{Internal walk voting objects}
\usage{
geo_map_swiss_voting_walk(data, date)
}
\arguments{
\item{data}{The data of the voting object}

\item{date}{The date of the vote}
}
\description{
Internal walk voting objects
}
\keyword{internal}
