% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_swiss.R
\name{geo_map_swiss_voting_write}
\alias{geo_map_swiss_voting_write}
\title{Internal write voting maps}
\usage{
geo_map_swiss_voting_write(map, data, id, title, level, date)
}
\arguments{
\item{map}{The map data}

\item{data}{The voting data}

\item{id}{The id of the voting object}

\item{title}{The title of the voting object}

\item{level}{The geometry level for the filename}

\item{date}{The date of the vote}
}
\description{
Internal write voting maps
}
\keyword{internal}
