% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_options.R
\name{geo_options}
\alias{geo_options}
\title{Set or get options for my package}
\usage{
geo_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{language}}{(\code{character};fr) The value of language }
}
}

