% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_carto.R
\name{geo_project_world}
\alias{geo_project_world}
\title{Project simple world map}
\usage{
geo_project_world(crs)
}
\arguments{
\item{crs}{A valid CRS string or object}
}
\value{
A ggplot map
}
\description{
Project simple world map
}
\keyword{internal}
