% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_carto.R
\name{geo_project_world_base_setup}
\alias{geo_project_world_base_setup}
\title{Internal: return ggplot2 settings for world map projections}
\usage{
geo_project_world_base_setup()
}
\value{
A list of ggproto elements
}
\description{
Internal: return ggplot2 settings for world map projections
}
\keyword{internal}
