% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_carto.R
\name{geo_project_world_tissot}
\alias{geo_project_world_tissot}
\title{Projet world map with tissot matrices}
\usage{
geo_project_world_tissot(crs)
}
\arguments{
\item{crs}{A valid CRS string}
}
\value{
A ggplot object
}
\description{
Projet world map with tissot matrices
}
