% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_swisstools.R
\name{geo_swiss_canton_code}
\alias{geo_swiss_canton_code}
\title{Get Swiss canton abbreviations}
\usage{
geo_swiss_canton_code(canton_name)
}
\arguments{
\item{canton_name}{A vector of Swiss canton names}
}
\value{
A vector of Swiss canton abbreviations
}
\description{
Get Swiss canton abbreviations
}
