% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_swisstools.R
\name{geo_swiss_canton_id}
\alias{geo_swiss_canton_id}
\title{Get Swiss canton ids}
\usage{
geo_swiss_canton_id(canton_name)
}
\arguments{
\item{canton_name}{A vector of Swiss canton names}
}
\value{
A vector of Swiss canton FSO ids
}
\description{
Get Swiss canton ids
}
