% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_swissmap.R
\name{geo_swiss_lakes_geom}
\alias{geo_swiss_lakes_geom}
\title{Swiss lakes for maps}
\usage{
geo_swiss_lakes_geom(fill_color = "skyblue")
}
\arguments{
\item{fill_color}{A color for the lakes (default: skyblue)}
}
\value{
A ggplot 2 layer for Swiss lakes
}
\description{
Swiss lakes for maps
}
