% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_swissmap.R
\name{geo_swiss_relief_geom}
\alias{geo_swiss_relief_geom}
\title{Swiss relief background for maps}
\usage{
geo_swiss_relief_geom()
}
\value{
A list of ggplot 2 layers for Swiss relief backgroup
}
\description{
Swiss relief background for maps
}
