% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_swissmap.R
\name{geo_walk_voting_maps}
\alias{geo_walk_voting_maps}
\title{Internal walk voting objects}
\usage{
geo_walk_voting_maps(data, date)
}
\arguments{
\item{date}{The date of the vote}

\item{...}{The data of the voting object}
}
\description{
Internal walk voting objects
}
