% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_mapping.R
\name{geo_wb_world_map}
\alias{geo_wb_world_map}
\title{Create world maps for an indicator}
\usage{
geo_wb_world_map(indicator)
}
\arguments{
\item{indicator}{A list of parameters}
}
\value{
Nothing
}
\description{
Create world maps for an indicator
}
