% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_swissmap.R
\name{geo_write_voting_maps}
\alias{geo_write_voting_maps}
\title{Internal write voting maps}
\usage{
geo_write_voting_maps(map, data, id, title, level, date)
}
\arguments{
\item{map}{The map data}

\item{data}{The voting data}

\item{id}{The id of the voting object}

\item{title}{The title of the voting object}

\item{date}{The date of the vote}
}
\description{
Internal write voting maps
}
