% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographer-package.R
\docType{package}
\name{geographer-package}
\alias{geographer}
\alias{geographer-package}
\title{geographer: Geography Vizualisations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides function and objects to establish vizualisations for my Geography lessons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://geographer.nenuial.org}
  \item Report bugs at \url{http://github.com/nenuial/geographer/issues/}
}

}
\author{
\strong{Maintainer}: Pascal Burkhard \email{pascal.burkhard@gmail.com} (\href{https://orcid.org/0009-0001-5504-5084}{ORCID})

}
\keyword{internal}
