% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_countrycode.R
\name{get_country_fips}
\alias{get_country_fips}
\title{Get country FIPS code}
\usage{
get_country_fips(country_name)
}
\arguments{
\item{country_name}{}
}
\value{
A FIPS country code
}
\description{
Get country FIPS code
}
