% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idbr_population.R
\name{get_idb_pyramid_5y_data}
\alias{get_idb_pyramid_5y_data}
\title{Five year population pyramid data}
\usage{
get_idb_pyramid_5y_data(country, year)
}
\arguments{
\item{country}{A FIPS code for the country}

\item{year}{An integer with the year}
}
\value{
A dataframe with age, female, and male population
}
\description{
Five year population pyramid data
}
