% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_functions.R
\name{get_ncdc_city_list}
\alias{get_ncdc_city_list}
\title{Read NCDC city list from cachedir}
\usage{
get_ncdc_city_list()
}
\value{
A dataframe of NCDC locations (cities)
}
\description{
Read NCDC city list from cachedir
}
