% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_functions.R
\name{get_noaa_climate_data}
\alias{get_noaa_climate_data}
\title{Function: retrieve climate data from the rnoaa::ncdc NOAA API and compute averages over the
time period 1990-2019.}
\usage{
get_noaa_climate_data(station_id, data_type = "temperature")
}
\arguments{
\item{station_id}{The string for the station ID}

\item{data_type}{The type of data needed. Can be "temperature" or "precipitation".}
}
\value{
A vectore with climate data (precipitation or temperature)
}
\description{
Function: retrieve climate data from the rnoaa::ncdc NOAA API and compute averages over the
time period 1990-2019.
}
