% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_basemap.R
\name{gph_base_map}
\alias{gph_base_map}
\alias{gph_base_tmap}
\title{Get a base map for a country}
\usage{
gph_base_map(country)

gph_base_tmap(country)
}
\arguments{
\item{country}{Country to highlight}
}
\value{
A leaflet map with the country borders
}
\description{
This function renders a base map with the country
highlighted.
}
\examples{
# Using Leaflet
gph_base_map("Switzerland")
gph_base_map("United Kingdom")

# Using tmap (v4)
gph_base_tmap("Switzerland")
gph_base_tmap("United Kingdom")

}
