% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_basemap.R
\name{gph_basemap}
\alias{gph_basemap}
\title{Get a basemap for a country}
\usage{
gph_basemap(country)
}
\arguments{
\item{country}{The country to get the basemap for}
}
\value{
A leaflet map with the country borders
}
\description{
Get a basemap for a country
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: needs rnaturalearthhires
gph_basemap("France")
gph_basemap("Switzerland")
\dontshow{\}) # examplesIf}
}
