% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_carto.R
\name{gph_boundbox}
\alias{gph_boundbox}
\title{NaturalEarth Bound box for planet Earth}
\usage{
gph_boundbox()
}
\value{
A simple feature
}
\description{
NaturalEarth Bound box for planet Earth
}
\examples{
gph_boundbox()
}
