% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_plot_world.R
\name{gph_country_population_tree}
\alias{gph_country_population_tree}
\title{Crete a treemap plot of the world countries by population}
\usage{
gph_country_population_tree()
}
\value{
A highcharter plot
}
\description{
Crete a treemap plot of the world countries by population
}
\examples{
gph_country_population_tree()
}
