% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_demography.R
\name{gph_hc_pyramid_relative}
\alias{gph_hc_pyramid_relative}
\title{Create highcharter population pyramid with relative values}
\usage{
gph_hc_pyramid_relative(country, year)
}
\arguments{
\item{country}{A string with the country name}

\item{year}{An integer for the year}
}
\value{
A highcharter graph
}
\description{
Create highcharter population pyramid with relative values
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: need a valid IDB API key
gph_hc_pyramid_relative("Switzerland", 2020)
\dontshow{\}) # examplesIf}
}
