% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_swiss.R
\name{gph_highcharter_map_swiss_votes}
\alias{gph_highcharter_map_swiss_votes}
\title{Plot latest swiss votation using highcharts}
\usage{
gph_highcharter_map_swiss_votes(
  geolevel = c("canton", "district", "municipality"),
  votedates,
  id,
  language = "FR"
)
}
\arguments{
\item{geolevel}{One of "canton", "district" or "municipality"}

\item{votedates}{The date of the vote}

\item{id}{The vote id}

\item{language}{One of "DE", "FR", "IT" or "RM"}
}
\value{
A highcharter map
}
\description{
Plot latest swiss votation using highcharts
}
