% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_world.R
\name{gph_historical_world_map}
\alias{gph_historical_world_map}
\title{Return Historical World Map}
\usage{
gph_historical_world_map(date)
}
\arguments{
\item{date}{A date object}
}
\value{
An sf object
}
\description{
Return Historical World Map
}
\examples{
lubridate::ymd("1945-06-01") |>
  gph_historical_world_map() |>
  ggplot2::ggplot() +
  ggplot2::geom_sf()

}
