% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_demography.R
\name{gph_lexgram}
\alias{gph_lexgram}
\title{Create lexgram for country}
\usage{
gph_lexgram(country, theme = ggplot2::theme_minimal())
}
\arguments{
\item{country}{A string with the country name}

\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Provides basic ggplot2 graph with the following data on it :
\describe{
\item{lex}{the total life expectancy (black dotted)}
\item{lex_male}{the male life expectancy (blue)}
\item{ley_female}{the male life expectancy (red)}
}
}
