% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_swiss.R
\name{gph_map_swiss_relief}
\alias{gph_map_swiss_relief}
\title{Swiss relief background for maps}
\usage{
gph_map_swiss_relief()
}
\value{
A list of ggplot 2 layers for Swiss relief backgroup
}
\description{
Swiss relief background for maps
}
\examples{
ggplot2::ggplot() +
  gph_map_swiss_relief() +
  gph_map_swiss_lakes()
}
