% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_world.R
\name{gph_map_world}
\alias{gph_map_world}
\title{Generate world map}
\usage{
gph_map_world(df, title, year, unit, palette, caption)
}
\arguments{
\item{df}{A dataframe with data breaks in the cut column}

\item{title}{The title of the map}

\item{year}{The year of the map}

\item{unit}{The units}

\item{palette}{The legend colors}

\item{caption}{The data source}
}
\value{
A ggplot2
}
\description{
Generate world map
}
\keyword{internal}
