% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_mini_globe.R
\name{gph_mini_globe}
\alias{gph_mini_globe}
\title{Create a globe with the countries in the list highlighted}
\usage{
gph_mini_globe(countries)
}
\arguments{
\item{countries}{A vector of country names}
}
\value{
A ggplot map
}
\description{
Create a globe with the countries in the list highlighted
}
\examples{
gph_mini_globe("Switzerland")
}
