% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_mini_globe.R
\name{gph_mini_globe_render}
\alias{gph_mini_globe_render}
\title{Internal: render mini globe}
\usage{
gph_mini_globe_render(country_codes, coords)
}
\arguments{
\item{country_codes}{The list of countries to highlight}

\item{coords}{The coordinates}
}
\value{
A ggplot map
}
\description{
Internal: render mini globe
}
\keyword{internal}
