% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_carto.R
\name{gph_project_world_base_setup}
\alias{gph_project_world_base_setup}
\title{Return ggplot2 settings for world map projections}
\usage{
gph_project_world_base_setup()
}
\value{
A list of ggproto elements
}
\description{
Return ggplot2 settings for world map projections
}
\keyword{internal}
