% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_carto.R
\name{gph_project_world_tissot}
\alias{gph_project_world_tissot}
\title{Projet world map with Tissot matrices}
\usage{
gph_project_world_tissot(crs)
}
\arguments{
\item{crs}{A valid CRS string or object}
}
\value{
A ggplot object
}
\description{
Projet world map with Tissot matrices
}
\examples{
gph_project_world_tissot("+proj=eqearth")
}
