% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_demography.R
\name{gph_pyramid}
\alias{gph_pyramid}
\title{Create population pyramid}
\usage{
gph_pyramid(country, year, theme = ggplot2::theme_minimal())
}
\arguments{
\item{country}{A string with the country name}

\item{year}{An integer for the year}

\item{theme}{A ggplot2 theme}
}
\value{
A ggplot graph
}
\description{
Create population pyramid
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: needs a valid IDB API key
gph_pyramid("Switzerland", 2020)
\dontshow{\}) # examplesIf}
}
