% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_demography.R
\name{gph_pyramid_relative}
\alias{gph_pyramid_relative}
\title{Create relative population pyramid (5 year cohorts)}
\usage{
gph_pyramid_relative(country, year, theme = ggplot2::theme_minimal())
}
\arguments{
\item{country}{A string with the country name}

\item{year}{An integer for the year}

\item{theme}{A ggplot2 theme}
}
\value{
A ggplot graph
}
\description{
Create relative population pyramid (5 year cohorts)
}
\examples{
gph_pyramid_relative("Switzerland", 2020)
}
