% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_world.R
\name{gph_wb_data}
\alias{gph_wb_data}
\title{Fetch World Bank data}
\usage{
gph_wb_data(indicator)
}
\arguments{
\item{indicator}{A list of parameters}
}
\value{
A dataframe
}
\description{
Fetch World Bank data
}
\keyword{internal}
