% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_world.R
\name{gph_wb_world_map}
\alias{gph_wb_world_map}
\title{Create world maps for an indicator}
\usage{
gph_wb_world_map(indicator)
}
\arguments{
\item{indicator}{A list of lists.
Each list should contain:
\describe{
\item{code}{A named vector with the name of the indicator and it's World Bank code}
\item{operation}{An expression that can make operations based on the indicator names}
\item{title}{Title of the map}
\item{file}{Name of the file}
\item{palette}{List with parameters \emph{palette} for the name of the palette,
\emph{type} for the type of palette and \emph{dir} for the direction}
\item{breaks}{Breaks for cutting the data}
\item{center}{Where is the center of the data (for diverging palettes), -1 otherwise}
\item{unit}{The unit of the indicator}
\item{year}{A vector with each year wanted}
}}
}
\description{
Create world maps for an indicator
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: saves the plots on file
indicators <- list(
  list(code      = c("population" = "EN.POP.DNST"),
       operation = expr(population),
       title     = geotools::gtl_translate_enfr(
                     "Population density", "Densité de population"
                   ),
       file      = "density",
       palette   = list(palette = "viridis::viridis",
                        type = "cont", dir = -1),
       breaks    = c(10,25,50,75,100,150,300,1000),
       center    = -1,
       unit      = "pp/km²",
       years     = c(2019)
  )
)

gph_wb_world_map(indicators)
\dontshow{\}) # examplesIf}
}
