% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_chine.R
\name{oc_chine_carte_2018_sex_ratio_par_region}
\alias{oc_chine_carte_2018_sex_ratio_par_region}
\title{OC Chine: carte du rapport homme/femmes par régions}
\usage{
oc_chine_carte_2018_sex_ratio_par_region(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
Une carte du rapport du nombre d'hommes
par rapport aux femmes par régions pour 2018.
}
\examples{
oc_chine_carte_2018_sex_ratio_par_region()
}
