% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_chine.R
\name{oc_chine_hc_sex_ratio}
\alias{oc_chine_hc_sex_ratio}
\title{Sex-ratio à la naissance}
\usage{
oc_chine_hc_sex_ratio()
}
\value{
A highcharter plot
}
\description{
Sex-ratio à la naissance
}
\examples{
oc_chine_hc_sex_ratio()
}
