% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_ecocitoyennete.R
\name{oc_ecocitoyennete_energy_per_source}
\alias{oc_ecocitoyennete_energy_per_source}
\title{OC Ecocitoyenneté : consommation d'énergie par source}
\usage{
oc_ecocitoyennete_energy_per_source(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
OC Ecocitoyenneté : consommation d'énergie par source
}
