% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_abortion_rights}
\alias{oc_geo_au_feminin_carte_abortion_rights}
\title{Géographie au féminin: carte des droits
en matière d'avortement}
\usage{
oc_geo_au_feminin_carte_abortion_rights(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
Géographie au féminin: carte des droits
en matière d'avortement
}
