% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_afghanistan_independance}
\alias{oc_geo_au_feminin_carte_afghanistan_independance}
\title{OC Géo au féminin: carte de l'Afhanistan 1919}
\usage{
oc_geo_au_feminin_carte_afghanistan_independance(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
OC Géo au féminin: carte de l'Afhanistan 1919
}
\examples{
oc_geo_au_feminin_carte_afghanistan_independance()
}
