% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_base_carte_wpp}
\alias{oc_geo_au_feminin_carte_base_carte_wpp}
\title{Base plot for WPP map}
\usage{
oc_geo_au_feminin_carte_base_carte_wpp(data)
}
\arguments{
\item{data}{WPP Lex data}
}
\value{
A ggplot2 map
}
\description{
Base plot for WPP map
}
\keyword{internal}
