% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_base_donnees_wpp}
\alias{oc_geo_au_feminin_carte_base_donnees_wpp}
\title{Filter WPP Lex data for graph}
\usage{
oc_geo_au_feminin_carte_base_donnees_wpp(data, filter)
}
\arguments{
\item{data}{WPP Lex data to be filtered}

\item{filter}{Period filter}
}
\value{
A tibble
}
\description{
Filter WPP Lex data for graph
}
\keyword{internal}
