% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_esperance_de_vie_wpp2012_2005_a_2010}
\alias{oc_geo_au_feminin_carte_esperance_de_vie_wpp2012_2005_a_2010}
\title{OC Géo au féminin: carte de la différence d'espérance de vie}
\usage{
oc_geo_au_feminin_carte_esperance_de_vie_wpp2012_2005_a_2010(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
OC Géo au féminin: carte de la différence d'espérance de vie
}
\examples{
oc_geo_au_feminin_carte_esperance_de_vie_wpp2012_2005_a_2010()
}
