% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_hc_abortion_rights}
\alias{oc_geo_au_feminin_carte_hc_abortion_rights}
\title{Géographie au féminin: carte interactive des droits
en matière d'avortement}
\usage{
oc_geo_au_feminin_carte_hc_abortion_rights()
}
\value{
A highcharts map
}
\description{
Géographie au féminin: carte interactive des droits
en matière d'avortement
}
