% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_hc_femmes_parlement}
\alias{oc_geo_au_feminin_carte_hc_femmes_parlement}
\title{Géographie au féminin: carte interactive du nombre de femmes au parlement}
\usage{
oc_geo_au_feminin_carte_hc_femmes_parlement()
}
\value{
A highcharts map
}
\description{
Géographie au féminin: carte interactive du nombre de femmes au parlement
}
\examples{
oc_geo_au_feminin_carte_hc_femmes_parlement()
}
