% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_hc_gii_2021}
\alias{oc_geo_au_feminin_carte_hc_gii_2021}
\title{Géographie au féminin: carte interactive du GII 2021}
\usage{
oc_geo_au_feminin_carte_hc_gii_2021()
}
\value{
A highcharts map
}
\description{
Carte interactive du Gender Inequality Index 2021
}
\examples{
oc_geo_au_feminin_carte_hc_gii_2021()
}
