% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_hc_sex_ratio}
\alias{oc_geo_au_feminin_carte_hc_sex_ratio}
\title{Géographie au féminin: carte interactive du sex-ratio à la naissance}
\usage{
oc_geo_au_feminin_carte_hc_sex_ratio()
}
\value{
A highcharts map
}
\description{
Géographie au féminin: carte interactive du sex-ratio à la naissance
}
