% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_hc_sigi_2023}
\alias{oc_geo_au_feminin_carte_hc_sigi_2023}
\title{Géographie au féminin: carte interactive du SIGI 2023}
\usage{
oc_geo_au_feminin_carte_hc_sigi_2023()
}
\value{
A highcharts map
}
\description{
Carte interactive du Social Institutions and Gender Index 2023
}
\examples{
oc_geo_au_feminin_carte_hc_sigi_2023()
}
