% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_hc_taux_avortement}
\alias{oc_geo_au_feminin_carte_hc_taux_avortement}
\title{Géographie au féminin: carte intéractive du
taux d'avortement}
\usage{
oc_geo_au_feminin_carte_hc_taux_avortement()
}
\value{
A highcharts map
}
\description{
Géographie au féminin: carte intéractive du
taux d'avortement
}
\examples{
oc_geo_au_feminin_carte_hc_taux_avortement()
}
