% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_hc_viol_europe}
\alias{oc_geo_au_feminin_carte_hc_viol_europe}
\title{OC Géo au féminin: carte interactive du taux de viols
déclarés}
\usage{
oc_geo_au_feminin_carte_hc_viol_europe()
}
\value{
A highcharts map
}
\description{
OC Géo au féminin: carte interactive du taux de viols
déclarés
}
