% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_hc_wbl}
\alias{oc_geo_au_feminin_carte_hc_wbl}
\title{Géographie au féminin: carte interactive du WBL}
\usage{
oc_geo_au_feminin_carte_hc_wbl()
}
\value{
A highcharts map
}
\description{
Carte interactive du Women, Business and the Law index
}
\examples{
oc_geo_au_feminin_carte_hc_wbl()
}
