% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_mondiale_suffrage_feminin}
\alias{oc_geo_au_feminin_carte_mondiale_suffrage_feminin}
\title{OC Géo au féminin: carte mondiale du suffrage féminin}
\usage{
oc_geo_au_feminin_carte_mondiale_suffrage_feminin(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot 2 theme}
}
\value{
A ggiraph map
}
\description{
Carte interactive de l'accès au suffrage féminin par pays
}
