% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_sex_ratio}
\alias{oc_geo_au_feminin_carte_sex_ratio}
\title{OC Géo au féminin: carte du sex ratio à la naissance}
\usage{
oc_geo_au_feminin_carte_sex_ratio(year, theme = ggplot2::theme_minimal())
}
\arguments{
\item{year}{The year}

\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
OC Géo au féminin: carte du sex ratio à la naissance
}
\examples{
oc_geo_au_feminin_carte_sex_ratio(2015)
}
