% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_carte_viol_europe}
\alias{oc_geo_au_feminin_carte_viol_europe}
\title{OC Géo au féminin: carte du taux de viols
déclarés}
\usage{
oc_geo_au_feminin_carte_viol_europe(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
OC Géo au féminin: carte du taux de viols
déclarés
}
\examples{
oc_geo_au_feminin_carte_viol_europe()
}
