% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_cartes_afghanistan}
\alias{oc_geo_au_feminin_cartes_afghanistan}
\alias{oc_geo_au_feminin_carte_afghanistan_independance}
\alias{oc_geo_au_feminin_carte_afghanistan}
\title{Géopolitique de l'Afghanistan}
\usage{
oc_geo_au_feminin_carte_afghanistan_independance(
  theme = ggplot2::theme_minimal()
)

oc_geo_au_feminin_carte_afghanistan(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
Cartes qui permettent de retracer l'évolution géopolitique
de l'Afghanistan depuis 1912.
}
\details{
Carte disponibles:
\itemize{
\item 1912: \code{oc_geo_au_feminin_carte_afghanistan_independance()}
\item 1950: \code{oc_geo_au_feminin_carte_afghanistan()}
}
}
\examples{
oc_geo_au_feminin_carte_afghanistan_independance()
oc_geo_au_feminin_carte_afghanistan()
}
