% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_cartes_avortements}
\alias{oc_geo_au_feminin_cartes_avortements}
\alias{oc_geo_au_feminin_carte_hc_abortion_rights}
\alias{oc_geo_au_feminin_carte_abortion_rights}
\alias{oc_geo_au_feminin_carte_hc_taux_avortement}
\alias{oc_geo_au_feminin_carte_taux_avortement}
\title{Avortement}
\usage{
oc_geo_au_feminin_carte_hc_abortion_rights()

oc_geo_au_feminin_carte_abortion_rights(theme = ggplot2::theme_minimal())

oc_geo_au_feminin_carte_hc_taux_avortement()

oc_geo_au_feminin_carte_taux_avortement(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A map
}
\description{
Cartes des droits en matière d'avortement ainsi que
des taux d'avortement.
}
\examples{
oc_geo_au_feminin_carte_hc_abortion_rights()
oc_geo_au_feminin_carte_abortion_rights()
oc_geo_au_feminin_carte_hc_taux_avortement()
oc_geo_au_feminin_carte_taux_avortement()
}
