% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_cartes_esperance_de_vie}
\alias{oc_geo_au_feminin_cartes_esperance_de_vie}
\alias{oc_geo_au_feminin_carte_esperance_de_vie_wpp2012_2005_a_2010}
\alias{oc_geo_au_feminin_carte_esperance_de_vie_wpp2019_2005_a_2010}
\alias{oc_geo_au_feminin_carte_esperance_de_vie_wpp2019_2015_a_2020}
\title{Différence d'espérance de vie}
\usage{
oc_geo_au_feminin_carte_esperance_de_vie_wpp2012_2005_a_2010(
  theme = ggplot2::theme_minimal()
)

oc_geo_au_feminin_carte_esperance_de_vie_wpp2019_2005_a_2010(
  theme = ggplot2::theme_minimal()
)

oc_geo_au_feminin_carte_esperance_de_vie_wpp2019_2015_a_2020(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A map
}
\description{
Cartes pour mettre en évidence les différences
d'espérence de vie entre hommes et femmes.
}
\examples{
oc_geo_au_feminin_carte_esperance_de_vie_wpp2012_2005_a_2010()
oc_geo_au_feminin_carte_esperance_de_vie_wpp2019_2005_a_2010()
oc_geo_au_feminin_carte_esperance_de_vie_wpp2019_2015_a_2020()
}
