% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_cartes_inegalites}
\alias{oc_geo_au_feminin_cartes_inegalites}
\alias{oc_geo_au_feminin_carte_sigi_2023}
\alias{oc_geo_au_feminin_carte_hc_sigi_2023}
\alias{oc_geo_au_feminin_carte_hc_gii_2021}
\alias{oc_geo_au_feminin_carte_hc_wbl}
\title{Mesure d'inégalité}
\usage{
oc_geo_au_feminin_carte_sigi_2023(theme = ggplot2::theme_minimal())

oc_geo_au_feminin_carte_hc_sigi_2023()

oc_geo_au_feminin_carte_hc_gii_2021()

oc_geo_au_feminin_carte_hc_wbl()
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A map
}
\description{
Ces différentes cartes présente plusieurs indicateurs de
l'inégalité entre hommes et femmes.
}
\details{
Indicateurs à disposition:
\itemize{
\item SIGI 2023: \code{oc_geo_au_feminin_carte_sigi_2023()} et \code{oc_geo_au_feminin_carte_hc_sigi_2023()}
\item GII 2021: \code{oc_geo_au_feminin_carte_hc_gii_2021()}
\item WB WBL 2022: \code{oc_geo_au_feminin_carte_hc_wbl()}
}
}
\examples{
# Sigi 2023
oc_geo_au_feminin_carte_sigi_2023()
oc_geo_au_feminin_carte_hc_sigi_2023()

# GII 2021
oc_geo_au_feminin_carte_hc_gii_2021()

# WB WBL 2022
oc_geo_au_feminin_carte_hc_wbl()
}
