% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_cartes_sex_ratio}
\alias{oc_geo_au_feminin_cartes_sex_ratio}
\alias{oc_geo_au_feminin_carte_sex_ratio}
\alias{oc_geo_au_feminin_carte_hc_sex_ratio}
\title{Sex ratio à la naissance}
\usage{
oc_geo_au_feminin_carte_sex_ratio(year, theme = ggplot2::theme_minimal())

oc_geo_au_feminin_carte_hc_sex_ratio()
}
\arguments{
\item{year}{The year}

\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
Cartes du sex ratio à la naissance.
}
\examples{
oc_geo_au_feminin_carte_sex_ratio(2015)
oc_geo_au_feminin_carte_hc_sex_ratio()
}
