% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_cartes_viol_europe}
\alias{oc_geo_au_feminin_cartes_viol_europe}
\alias{oc_geo_au_feminin_carte_viol_europe}
\alias{oc_geo_au_feminin_carte_hc_viol_europe}
\title{Viols}
\usage{
oc_geo_au_feminin_carte_viol_europe(theme = ggplot2::theme_minimal())

oc_geo_au_feminin_carte_hc_viol_europe()
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A map
}
\description{
Cartes du taux de viols déclarés en Europe.
}
\examples{
oc_geo_au_feminin_carte_viol_europe()
oc_geo_au_feminin_carte_hc_viol_europe()
}
