% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graph_2019_violences_sexuelles_police}
\alias{oc_geo_au_feminin_graph_2019_violences_sexuelles_police}
\title{OC Géographie au féminin: Violences sexuelles
\itemize{
\item raisons de ne pas contacter la police
}}
\usage{
oc_geo_au_feminin_graph_2019_violences_sexuelles_police(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
OC Géographie au féminin: Violences sexuelles
\itemize{
\item raisons de ne pas contacter la police
}
}
\examples{
oc_geo_au_feminin_graph_2019_violences_sexuelles_police()
}
