% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graph_difference_salariale_brut_par_formation}
\alias{oc_geo_au_feminin_graph_difference_salariale_brut_par_formation}
\title{OC Géo au féminin: différences salariales brut en Suisse
par formation}
\usage{
oc_geo_au_feminin_graph_difference_salariale_brut_par_formation(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot
}
\description{
OC Géo au féminin: différences salariales brut en Suisse
par formation
}
\examples{
oc_geo_au_feminin_graph_difference_salariale_brut_par_formation()
}
