% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graph_difference_salariale_brut_par_position}
\alias{oc_geo_au_feminin_graph_difference_salariale_brut_par_position}
\title{OC Géo au féminin: différences salariales brut en Suisse
Par position}
\usage{
oc_geo_au_feminin_graph_difference_salariale_brut_par_position(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot
}
\description{
OC Géo au féminin: différences salariales brut en Suisse
Par position
}
