% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graph_esperances_de_vie}
\alias{oc_geo_au_feminin_graph_esperances_de_vie}
\title{OC Géographie au féminin: Espérance de vie
Graphiques de l'expérance de vie}
\usage{
oc_geo_au_feminin_graph_esperances_de_vie(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
OC Géographie au féminin: Espérance de vie
Graphiques de l'expérance de vie
}
