% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graph_programmation_16_24_ans}
\alias{oc_geo_au_feminin_graph_programmation_16_24_ans}
\title{OC Géo au féminin: graphique sur le taux de jeunes apprenant à coder}
\usage{
oc_geo_au_feminin_graph_programmation_16_24_ans(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
OC Géo au féminin: graphique sur le taux de jeunes apprenant à coder
}
