% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graph_proportion_parlements}
\alias{oc_geo_au_feminin_graph_proportion_parlements}
\title{OC Géo au féminin: représentation politique
Pourcentage de femmes dans les parlements nationaux par régions}
\usage{
oc_geo_au_feminin_graph_proportion_parlements(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
OC Géo au féminin: représentation politique
Pourcentage de femmes dans les parlements nationaux par régions
}
