% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graph_scolarisation_filles_afghanistan}
\alias{oc_geo_au_feminin_graph_scolarisation_filles_afghanistan}
\title{OC Géographie au féminin: graphique du taux de scolarisation des filles en Afghanistant}
\usage{
oc_geo_au_feminin_graph_scolarisation_filles_afghanistan(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
OC Géographie au féminin: graphique du taux de scolarisation des filles en Afghanistant
}
