% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graph_suisse_avortements_par_regions}
\alias{oc_geo_au_feminin_graph_suisse_avortements_par_regions}
\title{OC Géographie au féminin: taux d'avortement
en Suisse par régions}
\usage{
oc_geo_au_feminin_graph_suisse_avortements_par_regions(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
OC Géographie au féminin: taux d'avortement
en Suisse par régions
}
