% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graph_suisse_votations_avortement}
\alias{oc_geo_au_feminin_graph_suisse_votations_avortement}
\title{OC Géographie au féminin: les résultats de différentes
votations fédérale sur le sujet de l'avortement}
\usage{
oc_geo_au_feminin_graph_suisse_votations_avortement(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
OC Géographie au féminin: les résultats de différentes
votations fédérale sur le sujet de l'avortement
}
