% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graphs_2019_violences_sexuelles}
\alias{oc_geo_au_feminin_graphs_2019_violences_sexuelles}
\alias{oc_geo_au_feminin_graph_2019_violences_sexuelles_experiences}
\alias{oc_geo_au_feminin_graph_2019_violences_sexuelles_actes}
\alias{oc_geo_au_feminin_graph_2019_violences_sexuelles_police}
\title{Expérience de violences sexuelles}
\usage{
oc_geo_au_feminin_graph_2019_violences_sexuelles_experiences(
  theme = ggplot2::theme_minimal()
)

oc_geo_au_feminin_graph_2019_violences_sexuelles_actes(
  theme = ggplot2::theme_minimal()
)

oc_geo_au_feminin_graph_2019_violences_sexuelles_police(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A graph
}
\description{
Graphiques visualisant les résultats de l'enquête sur
les violences sexuelles en Suisse réalisée par GFS.Bern
sur demande d'Amnesty international en avril 2019.
}
\examples{
oc_geo_au_feminin_graph_2019_violences_sexuelles_experiences()
oc_geo_au_feminin_graph_2019_violences_sexuelles_actes()
oc_geo_au_feminin_graph_2019_violences_sexuelles_police()
}
