% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graphs_difference_salariale}
\alias{oc_geo_au_feminin_graphs_difference_salariale}
\alias{oc_geo_au_feminin_graph_difference_salariale_brut_par_formation}
\alias{oc_geo_au_feminin_graph_difference_salariale_pourcent_par_formation}
\alias{oc_geo_au_feminin_graph_difference_salariale_brut_par_position}
\alias{oc_geo_au_feminin_graph_difference_salariale_pourcent_par_position}
\title{Différences salariales en Suisse}
\usage{
oc_geo_au_feminin_graph_difference_salariale_brut_par_formation(
  theme = ggplot2::theme_minimal()
)

oc_geo_au_feminin_graph_difference_salariale_pourcent_par_formation(
  theme = ggplot2::theme_minimal()
)

oc_geo_au_feminin_graph_difference_salariale_brut_par_position(
  theme = ggplot2::theme_minimal()
)

oc_geo_au_feminin_graph_difference_salariale_pourcent_par_position(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A graph
}
\description{
Graphiques présentant les différences salariale en Suisse
par formation et par position.
}
\examples{
oc_geo_au_feminin_graph_difference_salariale_brut_par_formation()
oc_geo_au_feminin_graph_difference_salariale_brut_par_position()
oc_geo_au_feminin_graph_difference_salariale_pourcent_par_formation()
oc_geo_au_feminin_graph_difference_salariale_pourcent_par_position()
}
