% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graphs_esperances_de_vie}
\alias{oc_geo_au_feminin_graphs_esperances_de_vie}
\alias{oc_geo_au_feminin_graph_esperances_de_vie}
\alias{oc_geo_au_feminin_graph_difference_esperance_de_vie}
\title{Espérance de vie}
\usage{
oc_geo_au_feminin_graph_esperances_de_vie(theme = ggplot2::theme_minimal())

oc_geo_au_feminin_graph_difference_esperance_de_vie(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A graph
}
\description{
Graphiques sur l'évolution de l'espérence de vie.
}
\examples{
oc_geo_au_feminin_graph_esperances_de_vie()
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: need credentials for HMD database
oc_geo_au_feminin_graph_difference_esperance_de_vie()
\dontshow{\}) # examplesIf}
}
