% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graphs_scolarisation_filles_afghanistan}
\alias{oc_geo_au_feminin_graphs_scolarisation_filles_afghanistan}
\alias{oc_geo_au_feminin_graph_scolarisation_filles_afghanistan}
\alias{oc_geo_au_feminin_hc_scolarisation_filles_afghanistan}
\title{Scolarisation des filles en Afghanistan}
\usage{
oc_geo_au_feminin_graph_scolarisation_filles_afghanistan(
  theme = ggplot2::theme_minimal(),
  line_color = "black"
)

oc_geo_au_feminin_hc_scolarisation_filles_afghanistan()
}
\arguments{
\item{theme}{A ggplot2 theme}

\item{line_color}{Graph line color}
}
\value{
A graph
}
\description{
Scolarisation des filles en Afghanistan
}
\examples{
oc_geo_au_feminin_graph_scolarisation_filles_afghanistan()
oc_geo_au_feminin_hc_scolarisation_filles_afghanistan()
}
