% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graphs_situation_eco_afghanistan}
\alias{oc_geo_au_feminin_graphs_situation_eco_afghanistan}
\alias{oc_geo_au_feminin_graph_situation_eco_afghanistan}
\alias{oc_geo_au_feminin_hc_situation_eco_afghanistan}
\title{PIB de l'Afghanistan par habitant}
\usage{
oc_geo_au_feminin_graph_situation_eco_afghanistan(
  theme = ggplot2::theme_minimal(),
  line_color = "black"
)

oc_geo_au_feminin_hc_situation_eco_afghanistan()
}
\arguments{
\item{theme}{A ggplot2 theme}

\item{line_color}{Graph line color}
}
\value{
A graph
}
\description{
Graphiques avec l'évolution du PIB par habitant en
Afghanistan
}
\examples{
oc_geo_au_feminin_graph_situation_eco_afghanistan()
oc_geo_au_feminin_hc_situation_eco_afghanistan()
}
