% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graphs_suisse_avortements}
\alias{oc_geo_au_feminin_graphs_suisse_avortements}
\alias{oc_geo_au_feminin_graph_suisse_avortements}
\alias{oc_geo_au_feminin_hc_suisse_avortements_par_ages}
\alias{oc_geo_au_feminin_graph_suisse_avortements_par_ages}
\alias{oc_geo_au_feminin_hc_suisse_avortements_par_regions}
\alias{oc_geo_au_feminin_graph_suisse_avortements_par_regions}
\alias{oc_geo_au_feminin_hc_suisse_votations_avortement}
\alias{oc_geo_au_feminin_graph_suisse_votations_avortement}
\title{Avortement en Suisse}
\usage{
oc_geo_au_feminin_graph_suisse_avortements(theme = ggplot2::theme_minimal())

oc_geo_au_feminin_hc_suisse_avortements_par_ages()

oc_geo_au_feminin_graph_suisse_avortements_par_ages(
  theme = ggplot2::theme_minimal()
)

oc_geo_au_feminin_hc_suisse_avortements_par_regions()

oc_geo_au_feminin_graph_suisse_avortements_par_regions(
  theme = ggplot2::theme_minimal()
)

oc_geo_au_feminin_hc_suisse_votations_avortement()

oc_geo_au_feminin_graph_suisse_votations_avortement(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A graph
}
\description{
Graphiques sur l'évolution du taux d'avortement en Suisse
et l'évolution politique à ce sujet.
}
\examples{
oc_geo_au_feminin_graph_suisse_avortements()
oc_geo_au_feminin_graph_suisse_avortements_par_ages()
oc_geo_au_feminin_hc_suisse_avortements_par_ages()
oc_geo_au_feminin_graph_suisse_avortements_par_regions()
oc_geo_au_feminin_hc_suisse_avortements_par_regions()
oc_geo_au_feminin_graph_suisse_votations_avortement()
oc_geo_au_feminin_hc_suisse_votations_avortement()
}
