% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_hc_programmation_16_24_ans}
\alias{oc_geo_au_feminin_hc_programmation_16_24_ans}
\title{OC Géo au féminin: graphique interactif sur le taux de jeunes apprenant à coder}
\usage{
oc_geo_au_feminin_hc_programmation_16_24_ans()
}
\value{
A highcharts graph
}
\description{
OC Géo au féminin: graphique interactif sur le taux de jeunes apprenant à coder
}
