% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_hc_scolarisation_filles_afghanistan}
\alias{oc_geo_au_feminin_hc_scolarisation_filles_afghanistan}
\title{OC Géographie au féminin: graphique du taux de scolarisation des filles en Afghanistan}
\usage{
oc_geo_au_feminin_hc_scolarisation_filles_afghanistan()
}
\value{
A highcharts graph
}
\description{
OC Géographie au féminin: graphique du taux de scolarisation des filles en Afghanistan
}
\examples{
oc_geo_au_feminin_hc_scolarisation_filles_afghanistan()
}
