% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_hc_situation_eco_afghanistan}
\alias{oc_geo_au_feminin_hc_situation_eco_afghanistan}
\title{OC Géo au féminin: représentation politique
Pourcentage de femmes dans les parlements nationaux par régions}
\usage{
oc_geo_au_feminin_hc_situation_eco_afghanistan()
}
\value{
A highcharts graph
}
\description{
OC Géo au féminin: représentation politique
Pourcentage de femmes dans les parlements nationaux par régions
}
\examples{
oc_geo_au_feminin_hc_situation_eco_afghanistan()
}
