% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_hc_suisse_avortements_par_ages}
\alias{oc_geo_au_feminin_hc_suisse_avortements_par_ages}
\title{OC Géographie au féminin: taux d'avortement en Suisse
par classe d'âge}
\usage{
oc_geo_au_feminin_hc_suisse_avortements_par_ages()
}
\value{
A hichgcharts graph
}
\description{
OC Géographie au féminin: taux d'avortement en Suisse
par classe d'âge
}
