% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_hc_suisse_avortements_par_regions}
\alias{oc_geo_au_feminin_hc_suisse_avortements_par_regions}
\title{OC Géographie au féminin: taux d'avortement
en Suisse par régions}
\usage{
oc_geo_au_feminin_hc_suisse_avortements_par_regions()
}
\value{
A highcharts
}
\description{
OC Géographie au féminin: taux d'avortement
en Suisse par régions
}
\examples{
oc_geo_au_feminin_hc_suisse_avortements_par_regions()
}
