% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_auf_feminin_graph_2019_violences_sexuelles_actes}
\alias{oc_geo_auf_feminin_graph_2019_violences_sexuelles_actes}
\title{OC Géographie au féminin: Actes de violences sexuelles}
\usage{
oc_geo_auf_feminin_graph_2019_violences_sexuelles_actes(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
OC Géographie au féminin: Actes de violences sexuelles
}
