% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geographie_au_feminin_carte_hc_contraception}
\alias{oc_geographie_au_feminin_carte_hc_contraception}
\title{Géographie au féminin: carte intéractive de
la prévalence de la contraception}
\usage{
oc_geographie_au_feminin_carte_hc_contraception()
}
\value{
A highcharts map
}
\description{
Géographie au féminin: carte intéractive de
la prévalence de la contraception
}
