% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_geo_au_feminin.R
\name{oc_geographie_au_feminin_carte_taux_avortement}
\alias{oc_geographie_au_feminin_carte_taux_avortement}
\title{Géographie au féminin: carte du
taux d'avortement}
\usage{
oc_geographie_au_feminin_carte_taux_avortement(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
Géographie au féminin: carte du
taux d'avortement
}
