% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_israel_palestine.R
\name{oc_israel_palestine_graph_migration_israel}
\alias{oc_israel_palestine_graph_migration_israel}
\title{OC Israël-Palestine: graphique des migrations}
\usage{
oc_israel_palestine_graph_migration_israel(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphique de l'immigration en Isräel
de 1949 à 2017.
}
