% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_israel_palestine.R
\name{oc_israel_palestine_graphs_migration}
\alias{oc_israel_palestine_graphs_migration}
\alias{oc_israel_palestine_graph_migration_israel}
\alias{oc_israel_palestine_hc_migration_israel}
\alias{oc_israel_palestine_graph_population_nae}
\alias{oc_israel_palestine_hc_immigration_sioniste}
\title{Migrations}
\usage{
oc_israel_palestine_graph_migration_israel(theme = ggplot2::theme_minimal())

oc_israel_palestine_hc_migration_israel()

oc_israel_palestine_graph_population_nae(theme = ggplot2::theme_minimal())

oc_israel_palestine_hc_immigration_sioniste()
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A graph
}
\description{
Graphiques sur le thème de l'immigration
en Israël.
}
\examples{
oc_israel_palestine_graph_migration_israel()
oc_israel_palestine_hc_migration_israel()
oc_israel_palestine_graph_population_nae()
oc_israel_palestine_hc_immigration_sioniste()
}
