% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_israel_palestine.R
\name{oc_israel_palestine_hc_immigration_sioniste}
\alias{oc_israel_palestine_hc_immigration_sioniste}
\title{OC Israël-Palestine: graphique dynamique
de l'immigration juive en Palestine
avant 1948}
\usage{
oc_israel_palestine_hc_immigration_sioniste()
}
\value{
A highcharts graph
}
\description{
OC Israël-Palestine: graphique dynamique
de l'immigration juive en Palestine
avant 1948
}
