% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_israel_palestine.R
\name{oc_israel_palestine_hc_migration_israel}
\alias{oc_israel_palestine_hc_migration_israel}
\title{OC Israël-Palestine: graphique dynamique des migrations}
\usage{
oc_israel_palestine_hc_migration_israel()
}
\value{
A highcharts graph
}
\description{
Graphique de l'immigration en Israël
de 1949 à 2017.
}
\examples{
oc_israel_palestine_hc_migration_israel()
}
