% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_israel_palestine.R
\name{oc_israel_palestine_population_nae}
\alias{oc_israel_palestine_population_nae}
\title{OC Israël-Palestine: graphique de la
population néé à l'étranger}
\usage{
oc_israel_palestine_population_nae(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphique avec la population néé à l'étranger
par pays de naissance d'après le recensement 2018.
}
