% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_russie.R
\name{oc_russie_carte_accroissement}
\alias{oc_russie_carte_accroissement}
\title{OC Russie: carte de l'accroissement naturel}
\usage{
oc_russie_carte_accroissement(
  theme = ggplot2::theme_minimal(),
  barwidth = 40,
  greyscale = F
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
Une carte de l'accroissement naturel en 2019
au niveau régional en Russie.
}
\concept{oc russie mortalité}
