% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_russie.R
\name{oc_russie_carte_adhesion_otan}
\alias{oc_russie_carte_adhesion_otan}
\alias{oc_russie_carte_hc_adhesion_otan}
\title{Adhésion à l'OTAN}
\usage{
oc_russie_carte_adhesion_otan(theme = ggplot2::theme_minimal())

oc_russie_carte_hc_adhesion_otan()
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
Carte de l'adhésion à l'OTAN par années
}
\examples{
oc_russie_carte_adhesion_otan()
oc_russie_carte_hc_adhesion_otan()
}
