% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_russie.R
\name{oc_russie_carte_divorces}
\alias{oc_russie_carte_divorces}
\title{OC Russie: carte du taux de divorce}
\usage{
oc_russie_carte_divorces(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
Une carte des divorces pour 1000 habitants
par régions en 2019.
}
\examples{
oc_russie_carte_divorces()
}
