% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_russie.R
\name{oc_russie_carte_evolution_population}
\alias{oc_russie_carte_evolution_population}
\title{OC Russie: carte de l'évolution démographique}
\usage{
oc_russie_carte_evolution_population(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
Une carte de l'évolution démographique Russe
entre 1990 et 2020.
}
