% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_russie.R
\name{oc_russie_carte_hc_elections_ukraine}
\alias{oc_russie_carte_hc_elections_ukraine}
\alias{oc_russie_carte_hc_elections_ukraine_2010}
\alias{oc_russie_carte_hc_elections_ukraine_2014}
\alias{oc_russie_carte_hc_elections_ukraine_2019}
\title{Cartes des élections présidentielles ukrainiennes}
\usage{
oc_russie_carte_hc_elections_ukraine_2010()

oc_russie_carte_hc_elections_ukraine_2014()

oc_russie_carte_hc_elections_ukraine_2019()
}
\value{
A higcharts map
}
\description{
Cartes des élections présidentielles ukrainiennes
en 2010, 2014 et 2019.
}
\examples{
oc_russie_carte_hc_elections_ukraine_2010()
oc_russie_carte_hc_elections_ukraine_2014()
oc_russie_carte_hc_elections_ukraine_2019()
}
