% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_russie.R
\name{oc_russie_carte_hc_religion}
\alias{oc_russie_carte_hc_religion}
\title{Carte de la religion majoritaire en 2012}
\usage{
oc_russie_carte_hc_religion()
}
\value{
A higcharts map
}
\description{
Religion avec le plus de pourcentage en 2012
par région.
}
\examples{
oc_russie_carte_hc_religion()
}
