% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_russie.R
\name{oc_russie_carte_natalite}
\alias{oc_russie_carte_natalite}
\title{OC Russie: carte du taux de natalité}
\usage{
oc_russie_carte_natalite(theme = ggplot2::theme_minimal(), barwidth = 40)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
Une carte du taux de natalité en 2019
au niveau régional en Russie.
}
\concept{oc russie fertilité}
