% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_attitude_eu}
\alias{oc_russie_graph_attitude_eu}
\title{OC Russie: Attitude envers l'UE}
\usage{
oc_russie_graph_attitude_eu(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
OC Russie: Attitude envers l'UE
}
\examples{
oc_russie_graph_attitude_eu()
}
