% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_attitude_us}
\alias{oc_russie_graph_attitude_us}
\title{OC Russie: Attitude envers les USA}
\usage{
oc_russie_graph_attitude_us(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
OC Russie: Attitude envers les USA
}
\examples{
oc_russie_graph_attitude_us()
}
