% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_demo_exa}
\alias{oc_russie_graph_demo_exa}
\title{OC Russie: demogram pour l'exa 2022}
\usage{
oc_russie_graph_demo_exa(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphique de l'évolution démographique de la Russie
entre 1980 et 2020. Graphique destiné à l'examen de
maturité2022.
}
\examples{
oc_russie_graph_demo_exa()
}
