% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_esperance_65_femmes}
\alias{oc_russie_graph_esperance_65_femmes}
\title{OC Russie: graphique de l'espérance de vie des femmes de 65 ans}
\usage{
oc_russie_graph_esperance_65_femmes(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphique de l'évolution de l'espérance de vie
de la Russie comparé à l'Italie et le Portugal
pour les femmes de 65 ans.
}
