% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_esperance_europe}
\alias{oc_russie_graph_esperance_europe}
\title{OC Russie: graphique de l'espérance de vie}
\usage{
oc_russie_graph_esperance_europe(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphique de l'évolution de l'espérance de vie
de la Russie comparé au reste de l'Europe.
}
