% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_immigration}
\alias{oc_russie_graph_immigration}
\title{OC Russie: graphique de l'immigration}
\usage{
oc_russie_graph_immigration(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphique de l'immigration en Russie avec les 5 pays
les plus représentés et une catégorie \code{autres}.
}
\examples{
oc_russie_graph_immigration()
}
