% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_mariages_et_divorces}
\alias{oc_russie_graph_mariages_et_divorces}
\title{Mariages et divorces}
\usage{
oc_russie_graph_mariages_et_divorces(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphiques des mariages et divorces en Russie
}
\examples{
oc_russie_graph_mariages_et_divorces()
}
