% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_naissance_et_avortements}
\alias{oc_russie_graph_naissance_et_avortements}
\title{OC Russie: graphique des naissances et avortements}
\usage{
oc_russie_graph_naissance_et_avortements(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphique des naissances en comparaison aux avortements.
}
\examples{
oc_russie_graph_naissance_et_avortements()
}
