% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_naissances}
\alias{oc_russie_graph_naissances}
\title{OC Russie: graphique des naissances}
\usage{
oc_russie_graph_naissances(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphique des naissances en Russie
jusqu'en 2017 (pour alignement avec avortements).
}
