% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_solde_migratoire}
\alias{oc_russie_graph_solde_migratoire}
\title{OC Russie: graphique du solde migratoire}
\usage{
oc_russie_graph_solde_migratoire(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphique du solde migratoire
en Russie par décénies.
}
