% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graphs_attitude}
\alias{oc_russie_graphs_attitude}
\alias{oc_russie_graph_attitude_us}
\alias{oc_russie_graph_attitude_eu}
\title{Attitude envers l'étranger}
\usage{
oc_russie_graph_attitude_us(theme = ggplot2::theme_minimal())

oc_russie_graph_attitude_eu(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphiques avec l'évolution de l'attitude des russes
envers l'étranger.
}
\examples{
oc_russie_graph_attitude_us()
oc_russie_graph_attitude_eu()
}
