% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graphs_esperance_de_vie}
\alias{oc_russie_graphs_esperance_de_vie}
\alias{oc_russie_graph_esperance_europe}
\alias{oc_russie_hc_esperance_europe}
\alias{oc_russie_graph_esperance_65_femmes}
\alias{oc_russie_graph_esperance_65_hommes}
\title{Espérance de vie russe}
\usage{
oc_russie_graph_esperance_europe(theme = ggplot2::theme_minimal())

oc_russie_hc_esperance_europe()

oc_russie_graph_esperance_65_femmes(theme = ggplot2::theme_minimal())

oc_russie_graph_esperance_65_hommes(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphiques qui montrent l'évolution de l'espérance de vie
de la Russie.
}
\examples{
oc_russie_graph_esperance_europe()
oc_russie_hc_esperance_europe()
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: needs credentials for HMD database
oc_russie_graph_esperance_65_femmes()
oc_russie_graph_esperance_65_hommes()
\dontshow{\}) # examplesIf}
}
