% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graphs_fertilite_europe}
\alias{oc_russie_graphs_fertilite_europe}
\alias{oc_russie_graph_fertilite_europe}
\alias{oc_russie_hc_fertilite_europe}
\title{Indice de fécondité}
\usage{
oc_russie_graph_fertilite_europe(theme = ggplot2::theme_minimal())

oc_russie_hc_fertilite_europe()
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph

A highcharts graph
}
\description{
Graphique de l'évolution de l'indice de fécondité
de la Russie comparé au reste de l'Europe.
}
\examples{
oc_russie_graph_fertilite_europe()
oc_russie_hc_fertilite_europe()
}
