% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_naissances}
\alias{oc_russie_graph_naissances}
\alias{oc_russie_graph_avortements}
\alias{oc_russie_graph_naissance_et_avortements}
\title{Naissances et avortements en Russie}
\usage{
oc_russie_graph_naissances(theme = ggplot2::theme_minimal())

oc_russie_graph_avortements(theme = ggplot2::theme_minimal())

oc_russie_graph_naissance_et_avortements(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphique des naissances et des avortements en Russie
jusqu'en 2017.
}
\examples{
oc_russie_graph_naissances()
oc_russie_graph_avortements()
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: needs patchwork to assemble the plot and only
# works if the package is loaded
oc_russie_graph_naissance_et_avortements()
\dontshow{\}) # examplesIf}
}
