% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_hc_ethnies}
\alias{oc_russie_hc_ethnies}
\title{Ethnies en Russie}
\usage{
oc_russie_hc_ethnies()
}
\value{
A highcharts graph
}
\description{
Graphique des 10 ethnies les plus fréquentes (hors Russes)
}
\examples{
oc_russie_hc_ethnies()
}
