% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_hc_pib_vs_suisse}
\alias{oc_russie_hc_pib_vs_suisse}
\title{PIB par habitant comparé entre la Russie et la Suisse}
\usage{
oc_russie_hc_pib_vs_suisse()
}
\value{
A highcharts graph
}
\description{
PIB par habitant comparé entre la Russie et la Suisse
}
\examples{
oc_russie_hc_pib_vs_suisse()
}
