% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_hc_religion}
\alias{oc_russie_hc_religion}
\title{Religion en Russie}
\source{
\url{https://www.levada.ru/en/2023/06/02/religious-beliefs/}
}
\usage{
oc_russie_hc_religion()
}
\value{
A graph
}
\description{
Religion en Russie
}
\examples{
oc_russie_hc_religion()
}
