% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_suisse.R
\docType{data}
\name{oc_suisse_1970_carte_suffrage_feminin_valais}
\alias{oc_suisse_1970_carte_suffrage_feminin_valais}
\title{OC Suisse: carte du vote sur le suffrage féminin au valais}
\format{
An object of class \code{function} of length 1.
}
\usage{
oc_suisse_1970_carte_suffrage_feminin_valais
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
ggplot2 map
}
\description{
Une carte du résultat de la votation sur le suffrage féminin
au canton du Valais le 12 avril 1970.
}
\concept{oc suisse votations suffrage femmes valais}
\keyword{datasets}
