% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_suisse.R
\name{oc_suisse_2021_graph_immigration_italienne}
\alias{oc_suisse_2021_graph_immigration_italienne}
\title{OC Suisse: graphique de l'immigration italienne}
\usage{
oc_suisse_2021_graph_immigration_italienne(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
ggplot2 graph
}
\description{
Une carte de l'introduction
}
\concept{oc suisse population immigration}
