% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_suisse.R
\name{oc_suisse_carte_1970_suffrage_feminin_valais}
\alias{oc_suisse_carte_1970_suffrage_feminin_valais}
\title{OC Suisse: carte du vote sur le suffrage féminin au valais}
\usage{
oc_suisse_carte_1970_suffrage_feminin_valais(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
ggplot2 map
}
\description{
Une carte du résultat de la votation sur le suffrage féminin
au canton du Valais le 12 avril 1970.
}
\examples{
oc_suisse_carte_1970_suffrage_feminin_valais()
}
\concept{oc suisse votations suffrage femmes valais}
