% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_suisse.R
\name{oc_suisse_carte_hc_2020_religion}
\alias{oc_suisse_carte_hc_2020_religion}
\title{Religion par commune}
\usage{
oc_suisse_carte_hc_2020_religion()
}
\value{
A highcharts map
}
\description{
Religion par commune
}
\examples{
oc_suisse_carte_hc_2020_religion()
}
