% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_suisse.R
\name{oc_suisse_carte_suffrage_feminin}
\alias{oc_suisse_carte_suffrage_feminin}
\title{Suffrage féminin}
\usage{
oc_suisse_carte_suffrage_feminin(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
ggplot2 map
}
\description{
Une carte de l'introduction du suffrage féminin
au niveau cantonal en Suisse.
}
\examples{
oc_suisse_carte_suffrage_feminin()
}
\concept{oc suisse votations suffrage femmes}
