% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_suisse.R
\name{oc_suisse_graph_2021_immigration_italienne}
\alias{oc_suisse_graph_2021_immigration_italienne}
\title{Immigration italienne en Suisse}
\usage{
oc_suisse_graph_2021_immigration_italienne(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Un graphique avec l'évolution de l'immigration
italienne en Suisse de 1981 jusqu'en 2021.
}
\examples{
oc_suisse_graph_2021_immigration_italienne()
}
