% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_suisse.R
\name{oc_suisse_hc_2023_nationalite}
\alias{oc_suisse_hc_2023_nationalite}
\title{Nationalité en Suisse en 2023}
\usage{
oc_suisse_hc_2023_nationalite()
}
\value{
A highcharts graph
}
\description{
Graphique des principales nationalité en Suisse en 2023
hors Suisses.
}
\examples{
oc_suisse_hc_2023_nationalite()
}
