% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translator}
\alias{translator}
\title{Creates translation function for named string list}
\usage{
translator(dictionary)
}
\arguments{
\item{dictionary}{A list of translations}
}
\value{
A translated string
}
\description{
Creates translation function for named string list
}
