% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_functions.R
\name{update_ncdc_city_list}
\alias{update_ncdc_city_list}
\title{Function: update the list of cities available for climate data, compute the
corresponding country column and store the data locally.}
\usage{
update_ncdc_city_list()
}
\description{
Function: update the list of cities available for climate data, compute the
corresponding country column and store the data locally.
}
