% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_cartography.R
\name{df_cartography_robinson_outline}
\alias{df_cartography_robinson_outline}
\title{Robinson outline}
\usage{
df_cartography_robinson_outline()
}
\value{
A sf object
}
\description{
Robinson special outline shape
}
\keyword{internal}
