% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_mapping.R
\name{geo_wb_data}
\alias{geo_wb_data}
\title{Fetch World Bank data}
\usage{
geo_wb_data(indicator)
}
\arguments{
\item{indicator}{A list of parameters}
}
\value{
A datafram
}
\description{
Fetch World Bank data
}
