% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_demography.R
\name{gph_demogram}
\alias{gph_demogram}
\title{Create demogram for country}
\usage{
gph_demogram(country, theme = ggplot2::theme_minimal())
}
\arguments{
\item{country}{A string with the country name}

\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Provides basic ggplot2 graph with the following data on it :
\describe{
\item{population}{the population curve (blue)}
\item{birth_rate}{the birth rate curve (grey)}
\item{death_rate}{the death_rate curve (black)}
}
}
