% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_plot_population.R
\name{gph_highcharter_pyramid}
\alias{gph_highcharter_pyramid}
\title{Create highcharter population pyramid}
\usage{
gph_highcharter_pyramid(country, year)
}
\arguments{
\item{country}{A string with the country name}

\item{year}{An integer for the year}
}
\value{
A highcharter graph
}
\description{
Create highcharter population pyramid
}
