% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_map_swiss.R
\name{gph_map_swiss_votes}
\alias{gph_map_swiss_votes}
\title{Plot latest swiss votation}
\usage{
gph_map_swiss_votes(
  geolevel = c("canton", "district", "municipality"),
  votedates,
  language = "FR",
  id
)
}
\arguments{
\item{geolevel}{One of "canton", "district" or "municipality"}

\item{votedates}{The date of the vote}

\item{language}{One of "DE", "FR", "IT" or "RM"}

\item{id}{The vote id}
}
\value{
A ggplot2 map
}
\description{
Plot latest swiss votation
}
