% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_geo_au_feminin.R
\name{oc_geo_au_feminin_graph_difference_esperance_de_vie}
\alias{oc_geo_au_feminin_graph_difference_esperance_de_vie}
\title{OC Géo au féminin: évolution de l'écart d'espérance de vie}
\usage{
oc_geo_au_feminin_graph_difference_esperance_de_vie(
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 graph
}
\description{
Graphique de l'évolution de l'écart entre
l'espérence de vie des hommes et des femmes
pour le Japon, la Suisse et les USA
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: need credentials for HMD database
oc_geo_au_feminin_graph_difference_esperance_de_vie()
\dontshow{\}) # examplesIf}
}
