% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_russie.R
\name{oc_russie_carte_fecondite}
\alias{oc_russie_carte_fecondite}
\alias{oc_russie_carte_hc_fecondite}
\title{Indice de fécondité en Russie}
\usage{
oc_russie_carte_fecondite(
  theme = ggplot2::theme_minimal(),
  barwidth = 40,
  greyscale = FALSE
)

oc_russie_carte_hc_fecondite()
}
\arguments{
\item{theme}{A ggplot2 theme}

\item{barwidth}{Width of legend bar}

\item{greyscale}{Boolean: whether to create a grey scale map}
}
\value{
A map
}
\description{
Une carte de l'indice de fécondité en 2023
au niveau régional en Russie.
}
\examples{
oc_russie_carte_fecondite()
oc_russie_carte_hc_fecondite()
}
\concept{oc russie fertilité}
