% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_russie.R
\name{oc_russie_carte_mariages}
\alias{oc_russie_carte_mariages}
\title{OC Russie: carte du taux de mariage}
\usage{
oc_russie_carte_mariages(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
A ggplot2 map
}
\description{
Une carte des mariages pour 1000 habitants
par régions en 2019.
}
\examples{
oc_russie_carte_mariages()
}
