% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oc_russie.R
\name{oc_russie_carte_mortalite}
\alias{oc_russie_carte_mortalite}
\title{Taux de mortalité en Russie}
\usage{
oc_russie_carte_mortalite(theme = ggplot2::theme_minimal(), barwidth = 40)
}
\arguments{
\item{theme}{A ggplot2 theme}

\item{barwidth}{Width of legend bar}
}
\value{
A ggplot2 map
}
\description{
Une carte du taux de mortalité en 2019
au niveau régional en Russie.
}
\examples{
oc_russie_carte_mortalite()
}
\concept{oc russie mortalité}
