% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_russie.R
\name{oc_russie_graph_depense_militaire}
\alias{oc_russie_graph_depense_militaire}
\title{Dépenses militaires}
\usage{
oc_russie_graph_depense_militaire(theme = ggplot2::theme_minimal())
}
\arguments{
\item{theme}{A ggplot2 theme}
}
\value{
An animated ggplot2
}
\description{
"Race plot" avec l'évolution des dépenses militaires
entre 2000 et 2020.
}
\examples{
oc_russie_graph_depense_militaire()
}
